backend/requirements.txt
# Hub Chantier - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client (pour tests et intégrations futures)
httpx>=0.26.0

# Development
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Linting (optionnel)
ruff>=0.1.0
